requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.0.2
certifi==2022.5.18.1
click==8.1.3
click-didyoumean==0.3.0
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.5
django-appconf==1.0.5
django-celery-email==3.0.0
django-phonenumber-field==6.2.0
django-phonenumbers==1.0.1
filelock==3.7.1
gunicorn==20.1.0
h11==0.13.0
kombu==5.2.4
packaging==21.3
phonenumbers==8.12.50
pipenv==2022.6.7
platformdirs==2.5.2
prompt-toolkit==3.0.29
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uvicorn==0.17.6
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
